+++ src/api/pharmacy.js
@@ -18,19 +18,17 @@ export const fetchMenuItems = async (lastFourDigits) => {
 	}
 };
 
-export const fetchItem = async ({ queryKey }) => {
+export const fetchItem = async (id) => {
 	try {
 		const { data, error } = await supabase
 			.from("pharmacy")
 			.select("id, place-name, address, phone-number, place-area, lat, lon")
-			.eq("id", queryKey[1])
+			.eq("id", id)
 			.single();
 
 		if (error) {
 			throw new Error(error.message);
 		}
-
-		console.log(data);
 		return data;
 	} catch (error) {
 		console.error("Error fetching item:", error.message);
