+++ src/components/Detail/Detail.jsx
@@ -12,7 +12,7 @@ export default function Detail() {
 		error
 	} = useQuery({
 		queryKey: ["pharmacy", detailId],
-		queryFn: fetchItem
+		queryFn: () => fetchItem(detailId)
 	});
 
 	if (isLoading) return <p>Loading...</p>;
