+++ src/pages/ListPage.jsx
@@ -1,9 +1,9 @@
 import { useQuery } from "@tanstack/react-query";
+import { useState } from "react";
 import { useLocation } from "react-router-dom";
+import { fetchMenuItems } from "../api/pharmacy";
 import ListPageInToggle from "../components/List/ListPageInToggle";
 import ListPageMap from "../components/List/ListPageMap";
-import { fetchMenuItems } from "../api/pharmacy";
-import { useState } from "react";
 
 function ListPage() {
 	const location = useLocation();
@@ -35,7 +35,6 @@ function ListPage() {
 				selectedMarkerId={selectedMarkerId}
 				setSelectedMarkerId={setSelectedMarkerId}
 			/>
-			<div>ListPage</div>
 		</>
 	);
 }
